chore(vendor): update @lightbasenl/backend (#3)
_This PR is created by sync and will be force-pushed daily. Overwriting
any manual changes done to this PR._


- fix(feature-flag): prevent invalid responses when starting a new api
with a different set of feature flags
We do rolling releases of all services. If a feature flag is added to
the new service, the response of the old service would already include
this new flag, erroring on strict response validation. This change makes
    sure that added or removed flags don't show up yet in the old api
    instance, based on the loaded feature flag config.
- chore(deps): bump to Compas v0.4.0

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>